+++ packages/calcite-components-react/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.7.0-next.12](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@1.7.0-next.11...@esri/calcite-components-react@1.7.0-next.12) (2023-08-29)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
+## [1.7.0-next.11](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@1.7.0-next.10...@esri/calcite-components-react@1.7.0-next.11) (2023-08-28)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
+## [1.7.0-next.10](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@1.7.0-next.9...@esri/calcite-components-react@1.7.0-next.10) (2023-08-28)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
 ## [1.7.0-next.9](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@1.7.0-next.8...@esri/calcite-components-react@1.7.0-next.9) (2023-08-26)
 
 **Note:** Version bump only for package @esri/calcite-components-react